+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.60.0] 2023-05-12
+
+### Added
+
+- upload plugin
+
+## [3.59.2] 2023-05-10
+
+### Fixed
+
+- persist responder meta through postbacks
+
 ## [3.59.1] 2023-05-03
 
 ### Fixed